Bullet updated.
